Emergency bug fix for message substitutions
[Your name] and [Your city] substitutions are not working while the
user fills out the form, so this workaround makes the substitution
during form submit.

In commit 5611bff (before we upgraded deps, upgraded Shadcn components,
migrated to Tailwind 4, and rewrote helptheducks.com in Next.js from
Vite) this bug took a lesser form. It would make one substitution when
the user first changed a form field, and then no further substitutions
to the message would be made even if the user did not edit the message.

After all the upgrades, just prior to this commit, substition is
broken entirely. All fields were always marked as dirty before any
user interaction.

I tested replacing the shadcn email field with
<input ...register("email") /> and the email field was no longer marked
as dirty before user interaction, so the issue seems to be related to
the shadcn components or the way they are wired up with react-hook-form.