fix: stop using --acl public-read
Checking if this fix will no longer require that buckets have ACLs
turned on. It seems to no longer be the default for buckets to have
ACLs enabled and is not recommended except when individual items need
varying permissions. Since we use CloudFront distributions to serve
the website, I'm hoping this will still work.